refactor-refactor post and comment code
